requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
google-genai==1.11.0
